+++ multiplay/skirmish/fb_includes/__tac_log_production.js
@@ -327,6 +327,28 @@ function produceHeavyCavalry(factory) {
 	});
 }
 
+function produceHeavyRepair(factory) {
+	// A cool idea from the 'Peacemaker' bot (by 'duckfood'). 
+	// With enough nearby repair turrets, it is possible to greatly to extend the lifespan of a frontline unit.
+	const repairTurrets = [
+		WEAPONS['Heavy Repair Turret']
+	].reverse();
+	
+	const heavyRepairPropulsion = [
+		PROPULSIONS["Wheels"], 
+		PROPULSIONS["Half-tracks"], 
+		PROPULSIONS["Tracks"]
+	].reverse();
+
+	return produceVehicle({
+		factory: factory, 
+		weaponList: repairTurrets, 
+		propulsionList: heavyRepairPropulsion,
+		maxBodyWeight: BODY_WEIGHT.HEAVY
+	});
+}
+
+
 function produceLandAPFireSupport(factory) {
 	// Order these by tech level if you want the most technologically advanced weapon to be used
 	const fireSupportWeapons = [
@@ -393,7 +415,7 @@ function produceHighVolumeAAUnit(factory) {
 		factory: factory, 
 		weaponList: shortRangeAAWeapons, 
 		propulsionList: airDefenceArtilleryPropulsions, 
-		maxBodyWeight: BODY_WEIGHT.MEDIUM
+		maxBodyWeight: BODY_WEIGHT.HEAVY
 	});
 }
 
@@ -475,9 +497,7 @@ function produceInfantry(factory) {
 }
 
 function produceLandUnitCategory(category, factory) {
-	let factoryInProduction = false;   
-
-	let r = Math.floor(Math.random() * 4);		// this should be one of the few (if any) Math.random() calls in FishBot.
+	let factoryInProduction = false;  
 	
 	switch (category) {
 		case 'heavyCavalry':
@@ -487,17 +507,16 @@ function produceLandUnitCategory(category, factory) {
 			factoryInProduction = factoryInProduction || produceLightCavalry(factory);
 			break;
 		case 'shortRangeArtillery':
-			factoryInProduction = factoryInProduction || produceLandFireSupportGeneric(factory);
+			factoryInProduction = factoryInProduction || produceLandAPFireSupport(factory);
 			break;
 		case 'ADA':
-			if (r === 0) {
-				factoryInProduction = factoryInProduction || produceAAFlakUnit(factory);
-			} else {
-				factoryInProduction = factoryInProduction || produceHighVolumeAAUnit(factory);
-			}
+			factoryInProduction = factoryInProduction || produceHighVolumeAAUnit(factory);
 			break;
 		case 'sensor': 
 			factoryInProduction = factoryInProduction || produceLandRecon(factory);
+			break;
+		case 'repair':
+			factoryInProduction = factoryInProduction || produceHeavyRepair(factory);
 			break;
 		default:
 			factoryInProduction = factoryInProduction || produceLightCavalry(factory);
@@ -551,6 +570,10 @@ function getDroidFbGroupClassification(droid) {
 		return DIVISION.AIR_DEFENCE_RESERVE;
 	}
 
+	if (flags & OBJ_FLAGS.REPAIR) {
+		return DIVISION.MAINTENANCE_RESERVE;
+	}
+ 
 	if (droid.droidType === DROID_SENSOR) {
 		// manually accessing the DroidObject properties as I have run out of bits in OBJ_FLAGS.
 		return DIVISION.SENSOR_RESERVE;